requirements.txt
alembic
Flask
requests
Flask-Cors
Flask-Migrate
Flask-SQLAlchemy
gunicorn
pillow
python-dotenv
SQLAlchemy
pandas
numpy
PyMySQL
bcrypt
Flask-Bcrypt
click
Flask-JWT-Extended
Flask-Mail
flask-marshmallow
marshmallow-sqlalchemy
marshmallow
cryptography